+++ apps/api/src/core/environment/dto/create-env.input.ts
@@ -1,0 +1,12 @@
+import { ApiProperty } from "@nestjs/swagger";
+import { IsString } from "class-validator";
+
+export class CreateEnvInput {
+  @ApiProperty()
+  @IsString()
+  name: string;
+
+  @ApiProperty()
+  @IsString()
+  projectId: string;
+}