+++ apps/ui/app/routes/action+/project+/create.ts
@@ -1,7 +1,7 @@
 import { ProjectDto } from "@montelo/browser-client";
 import { ActionFunction, json } from "@remix-run/node";
 import _ from "lodash";
-import { withAuth } from "../../../auth/withAuth";
+import { withAuth } from "~/auth/withAuth";
 
 export const action: ActionFunction = withAuth(async ({ api, request }) => {
   const formData = await request.formData();