+++ apps/log-server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @montelo/log-server
 
+## 1.1.59
+
+### Patch Changes
+
+- y
+- Updated dependencies
+  - @montelo/db@1.1.59
+
 ## 1.1.58
 
 ### Patch Changes